+++ conman/routes/models.py
@@ -4,7 +4,6 @@ from polymorphic.models import PolymorphicModel
 
 from .handlers import RouteViewHandler
 from .managers import RouteManager
-from .utils import import_from_dotted_path
 from .validators import (
     validate_end_in_slash,
     validate_no_dotty_subpaths,
@@ -49,7 +48,7 @@ class Route(PolymorphicModel, metaclass=UnboundViewMeta):
     )
 
     objects = RouteManager()
-    handler = RouteViewHandler.path()
+    handler_class = RouteViewHandler
 
     def __str__(self):
         """Display a Route's class and url."""
@@ -63,10 +62,6 @@ class Route(PolymorphicModel, metaclass=UnboundViewMeta):
         descendants = others.filter(url__startswith=self.url)
         return descendants.order_by('url')
 
-    def get_handler_class(self):
-        """Import a class from the python path string in `self.handler`."""
-        return import_from_dotted_path(self.handler)
-
     def get_handler(self):
         """
         Get an instance of the handler for this Route instance.
@@ -77,7 +72,7 @@ class Route(PolymorphicModel, metaclass=UnboundViewMeta):
         try:
             return self._handler
         except AttributeError:
-            self._handler = self.get_handler_class()(self)
+            self._handler = self.handler_class(self)
             return self._handler
 
     def handle(self, request, path):